requirements.txt
appdirs==1.4.4
backcall==0.1.0
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.1
cssselect==1.1.0
decorator==4.4.2
distlib==0.3.0
docutils==0.15.2
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
fuzzywuzzy==0.18.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
imutils==0.5.3
isort==4.3.21
jedi==0.17.0
jellyfish==0.8.2
jmespath==0.10.0
joblib==0.15.1
lxml==4.5.1
mccabe==0.6.1
oauthlib==3.1.0
opencv-python==4.2.0.34
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
protobuf==3.12.0
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==19.8.18
Pygments==2.6.1
pylint==2.3.1
PyMySQL==0.9.3
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.13.0.1.5
PyQtWebEngine==5.15.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-Levenshtein==0.12.0
pytz==2020.1
pywin32==227
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.5.14
requests==2.23.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.3
webencodings==0.5.1
wrapt==1.11.2